% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionaries.R
\docType{class}
\name{DictionaryLearners}
\alias{DictionaryLearners}
\title{Registered Learners}
\format{\code{\link{R6Class}} object}
\usage{
DictionaryLearners
}
\description{
\code{mlr.learners} is a \code{\link{Dictionary}} used to manage learners.
}
\examples{
mlr.learners$ids
mlr.learners$contains("classif.dummy")
mlr.learners$get("classif.dummy")
}
\keyword{datasets}
