% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionaries.R
\docType{class}
\name{DictionaryTasks}
\alias{DictionaryTasks}
\title{Registered Tasks}
\format{\code{\link{R6Class}} object}
\usage{
DictionaryTasks
}
\description{
\code{Tasks} is a \code{\link{Dictionary}} used to manage tasks.
}
\keyword{datasets}
