% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_learner.R
\name{expect_learner}
\alias{expect_learner}
\title{Check Learner}
\usage{
expect_learner(lrn)
}
\arguments{
\item{lrn}{(\code{\link[=Learner]{Learner()}})\cr
Object of type \code{\link[=Learner]{Learner()}}.}
}
\description{
Check if Learner object is valid.
}
