% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.learners.helper.R
\name{learnerArgsToControl}
\alias{learnerArgsToControl}
\title{Convert arguments to control structure.}
\usage{
learnerArgsToControl(control, ..., .defaults = list(),
  .restrict = FALSE)
}
\arguments{
\item{control}{\code{\link{function}}\cr
Function that creates control structure.}

\item{...}{\link{any}\cr
Arguments for control structure function.}

\item{.defaults}{\code{\link{list}}\cr
Initial default values that the control function is called on.
One can use this if control should be called on different defaults than in its signature.
The values in ... will overwrite these.
Default is empty list.}

\item{.restrict}{(\code{logical(1)})\cr
Restrict the elements from \code{...}, which are passed to \code{control}, only to the arg-names
that occur in the signature of \code{control}?
Default is \code{FALSE}.}
}
\value{
Control structure for learner.
}
\description{
Find all elements in \code{...} which are not missing and
call \code{control} on them.
}
