% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learners.R
\docType{class}
\name{Learners}
\alias{Learners}
\title{Registered Learners}
\format{\code{\link{R6Class}} object}
\usage{
Learners
}
\description{
\code{Learners} is a \code{\link{Dictionary}} used to manage learners.
}
\examples{
Learners$ids
Learners$contains("classif.dummy")
Learners$get("classif.dummy")
}
\keyword{datasets}
