% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\docType{class}
\name{Measures}
\alias{Measures}
\title{Registered Resampling Methods}
\format{\code{\link{R6Class}} object}
\usage{
Measures
}
\description{
\code{Measures} is a \code{\link{Dictionary}} used to manage performance measures.
}
\keyword{datasets}
