% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.learners.R
\docType{class}
\name{mlr.learners}
\alias{mlr.learners}
\title{Registered Learners}
\format{\code{\link[=R6Class]{R6Class()}} object}
\usage{
mlr.learners
}
\description{
\code{mlr.learners} is a \code{\link[=Dictionary]{Dictionary()}} used to manage learners.
}
\examples{
mlr.learners$ids
mlr.learners$contains("classif.dummy")
mlr.learners$get("classif.dummy")
}
\keyword{datasets}
