% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionaries.R
\docType{class}
\name{mlr.measures}
\alias{mlr.measures}
\title{Registered Resampling Methods}
\format{\code{\link{R6Class}} object}
\usage{
mlr.measures
}
\description{
\code{mlr.measures} is a \code{\link{Dictionary}} used to manage performance measures.
}
\keyword{datasets}
