% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\docType{class}
\name{mlr.tasks}
\alias{mlr.tasks}
\title{Registered Tasks}
\format{\code{\link{R6Class}} object}
\usage{
mlr.tasks
}
\description{
\code{Tasks} is a \code{\link{Dictionary}} used to manage tasks.
}
\keyword{datasets}
