% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MlrModel.R
\docType{data}
\name{MlrModel}
\alias{MlrModel}
\title{MlrModel}
\format{\code{\link{R6Class}} object}
\usage{
MlrModel
}
\description{
A \code{\link[R6]{R6Class}} containing data of a \code{\link{train}}.
}
\section{Fields}{

\describe{
\item{\code{task}}{[\code{\link{Task}}]: Task used to fit the model.}

\item{\code{learner}}{[\code{\link{Learner}}]: Learner used to fit the model.}

\item{\code{wrapped.model}}{[any]: Result of the model fit as returned by third party packages.}

\item{\code{train}}{[\code{integer}]: Indices of training data used to fit the model.}

\item{\code{log}}{[\code{\link{TrainLog}}]: Logging information from model fit.}
}}

\keyword{datasets}
