% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlrng-package}
\alias{mlrng}
\alias{mlrng-package}
\title{mlrng: Machine Learning in R - Next Generation}
\description{
Machine Learning in R. Next Gen.
}
\author{
\strong{Maintainer}: Bernd Bischl \email{bernd_bischl@gmx.de}

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com}
}

}
\keyword{internal}
