% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult.R
\docType{data}
\name{ResampleResult}
\alias{ResampleResult}
\title{ResampleResult}
\format{\code{\link{R6Class}} object}
\usage{
ResampleResult
}
\description{
A \code{\link[R6]{R6Class}} containing data of a \code{\link{resample}}.
}
\section{Fields}{

\describe{
\item{\code{aggr}}{[\code{named numeric}]: Aggregated performance measures.}

\item{\code{data}}{[\code{data.table}]: Data stored in a tabular format.}
}}

\keyword{datasets}
