% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resamplings.R
\docType{class}
\name{Resamplings}
\alias{Resamplings}
\title{Registered Resampling Methods}
\format{\code{\link{R6Class}} object}
\usage{
Resamplings
}
\description{
\code{Resamplings} is a \code{\link{Dictionary}} used to manage resampling methods.
}
\keyword{datasets}
