% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split.R
\docType{data}
\name{Split}
\alias{Split}
\title{Split}
\format{\code{\link{R6Class}} object}
\usage{
Split
}
\description{
A \code{\link[R6]{R6Class}} containing data of a split into training and test set.
}
\section{Fields}{

\describe{
\item{\code{train}}{[\code{integer}]: Training indices}

\item{\code{test}}{[\code{integer}]: Test indices}
}}

\keyword{datasets}
