% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr.R
\docType{data}
\name{TaskRegr}
\alias{TaskRegr}
\title{Regression Tasks}
\format{\code{\link{R6Class}} object}
\usage{
TaskRegr
}
\value{
[\code{\link{TaskRegr}}].
}
\description{
A \code{\link[R6]{R6Class}} to construct classification tasks.
}
\section{Fields}{

\describe{
\item{\code{id}}{[\code{character(1)}]: Identifier of the task.}

\item{\code{backend}}{[\code{Backend}]: Internal abstraction for data access. Leave this alone unless you know exactly what you are doing.}

\item{\code{get}}{[\code{function(rows = NULL, cols = NULL)}]: Names of the features.}

\item{\code{features}}{[\code{character}]: Names of the features.}

\item{\code{formula}}{[\code{character}]: Formula describing the learning task.}

\item{\code{target}}{[\code{character(1)}]: Name of the target column.}
}}

\examples{
task = TaskRegr$new(data = iris, target = "Sepal.Length")
task$formula
}
\seealso{
Other Tasks: \code{\link{TaskClassif}}
}
\keyword{datasets}
