% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tasks.R
\docType{class}
\name{Tasks}
\alias{Tasks}
\title{Registered Tasks}
\format{\code{\link{R6Class}} object}
\usage{
Tasks
}
\description{
\code{Tasks} is a \code{\link{Dictionary}} used to manage tasks.
}
\keyword{datasets}
