% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSupervised.R
\docType{data}
\name{TaskSupervised}
\alias{TaskSupervised}
\title{Base Class for Supervised Tasks}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaskSupervised
}
\value{
[\code{\link{TaskSupervised}}].
}
\description{
A \code{\link[R6]{R6Class}} to construct supervised tasks.
This is the abstract base class, do not use directly!
}
\section{Fields}{

\describe{
\item{\code{id}}{[\code{character(1)}]: Identifier of the task.}

\item{\code{features}}{[\code{character}]: Names of the features.}

\item{\code{formula}}{[\code{character}]: Formula describing the learning task.}

\item{\code{target}}{[\code{character(1)}]: Name of the target column.}
}}

\seealso{
Other Tasks: \code{\link{TaskClassif}},
  \code{\link{TaskRegr}}, \code{\link{Task}}
}
\keyword{datasets}
