% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainLog.R
\docType{data}
\name{TrainLog}
\alias{TrainLog}
\title{TrainLog}
\format{\code{\link[=R6Class]{R6Class()}} object}
\usage{
TrainLog
}
\description{
A \code{\link[R6:R6Class]{R6::R6Class()}} containing logging information of a \code{\link[=train]{train()}} call.
Constructor is called with output (\code{list}): List of output, messages, warnings and errors occured in training and the training time.
}
\section{Fields}{

\describe{
\item{\code{train.time}}{(\code{numeric(1)}): Training time.}

\item{\code{messages}}{(\link{list}): List of messages encountered while training the model.}

\item{\code{warnings}}{([list)]: List of warnings encountered while training the model.

[list)]: R:list)}

\item{\code{errors}}{([list)]: List of errors encountered while training the model.

[list)]: R:list)}

\item{\code{output}}{(\link{list}): Remaining output from model fit, only saved if \code{mlrng.keep.train.output} option is set to \code{TRUE}.}
}}

\keyword{datasets}
