% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View.R
\docType{data}
\name{View}
\alias{View}
\title{Abstract View on a Data Source}
\format{\code{\link{R6Class}} object}
\usage{
View
}
\value{
[\code{View}].
}
\description{
This class currently sits on top of \pkg{dplyr} and is intended to provide
a view on a read-only data source.
}
\section{Fields}{

\describe{
\item{\code{pars}}{[\code{list()}]:\cr
Arguments for \code{\link[DBI]{dbConnect}} to establish a connection and re-connect if
necessary.}

\item{\code{name}}{[\code{character(1)}]:\cr Name of the table in the data base.}

\item{\code{rowid.col}}{[\code{character(1)}]:\cr Name of a unique id column in the table.}

\item{\code{con}}{[\code{connection}]:\cr Connection to the data base.}

\item{\code{raw.tbl}}{[\code{src_sql}]:\cr \pkg{dplyr} source of the data base table.}

\item{\code{tbl}}{[\code{src_sql}]:\cr \pkg{dplyr} source of the data base table, pre-filtered
and pre-selected to the active view.}

\item{\code{data}}{[\code{function(rows = NULL, cols = NULL}]\cr
Function to extract rows and columns from the subset of active rows and active cols.}

\item{\code{active.rows}}{[\code{vector}]\cr
Subset of values in \code{rowid.col}. Can be overwritten to change the view.}

\item{\code{active.cols}}{[\code{vector}]\cr
Subset of column names. Can be overwritten to change the view.}

\item{\code{checksum}}{[\code{character(1)}]\cr
Hash of the active view.}

\item{\code{nrow}}{[\code{integer(1)}]\cr
Number of rows in active view.}

\item{\code{ncol}}{[\code{integer(1)}]\cr
Number of cols in active view.}

\item{\code{head}}{[\code{function(n = 6)}]\cr
Get the first \code{n} rows of the active view.}

\item{\code{types}}{[\code{named character}]\cr
Class information about all columns in the active view.}

\item{\code{na.cols}}{[\code{named integer}]\cr
Number of missing values per column in the active view.}

\item{\code{distinct}}{[\code{function(col)}]\cr
Levels of column \code{col} in the active view.}
}}

\keyword{datasets}
