% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrappedModel.R
\docType{data}
\name{WrappedModel}
\alias{WrappedModel}
\title{WrappedModel}
\format{\code{\link{R6Class}} object}
\usage{
WrappedModel
}
\description{
A \code{\link[R6]{R6Class}} containing data of a \code{\link{train}}.
}
\section{Fields}{

\describe{
\item{\code{task}}{[\code{\link{Task}}]: Task used to fit the model.}

\item{\code{learner}}{[\code{\link{Learner}}]: Learner used to fit the model.}

\item{\code{model}}{[any]: Result of the model fit as returned by third party packages.}

\item{\code{train}}{[\code{integer}]: Indices of training data used to fit the model.}
}}

\keyword{datasets}
