% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkResult.R
\docType{data}
\name{BenchmarkResult}
\alias{BenchmarkResult}
\title{BenchmarkResult}
\format{\code{\link{R6Class}} object}
\usage{
BenchmarkResult
}
\description{
A \code{\link[R6]{R6Class}} containing data of a \code{\link{benchmark}}.
}
\section{Fields}{

\describe{
\item{\code{data}}{[\code{data.table}]: Data stored in a tabular format.}
}}

\keyword{datasets}
