% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlrng-package}
\alias{mlrng}
\alias{mlrng-package}
\title{mlrng: Machine Learning in R - Next Generation}
\description{
For bug reports and feature requests please use the tracker:
\url{https://github.com/mlr-org/mlrng}.
Package options are convered in \link{mlrng-config}.
}
\author{
\strong{Maintainer}: Bernd Bischl \email{bernd_bischl@gmx.de}

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com}
}

}
