% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\docType{data}
\name{Task}
\alias{Task}
\title{Base Class for Tasks}
\format{\code{\link{R6Class}} object}
\usage{
Task
}
\value{
[\code{\link{Task}}].
}
\description{
A \code{\link[R6]{R6Class}} to construct tasks.
This is the abstract base class, do not use directly!
}
\section{Fields}{

\describe{
\item{\code{id}}{[\code{character(1)}]: Identifier of the task.}

\item{\code{features}}{[\code{character}]: Names of the features.}

\item{\code{formula}}{[\code{character}]: Formula describing the learning task.}
}}

\seealso{
Other Tasks: \code{\link{TaskClassif}},
  \code{\link{TaskRegr}}, \code{\link{TaskSupervised}}
}
\keyword{datasets}
