% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\docType{data}
\name{Task}
\alias{Task}
\title{Basic Tasks}
\format{\code{\link{R6Class}} object}
\usage{
Task
}
\value{
[\code{\link{Task}}].
}
\description{
This is the abstract base class for task objects.
Use \code{\link{TaskClassif}} or \code{\link{TaskRegr}} to construct tasks instead of this class.
}
\section{Fields}{

\describe{
\item{\code{id}}{[\code{character(1)}]: Identifier of the task.}

\item{\code{backend}}{[\code{Backend}]: Internal abstraction for data access. Leave this alone unless you know exactly what you are doing.}

\item{\code{get}}{[\code{function(rows = NULL, cols = NULL)}]: Names of the features.}

\item{\code{features}}{[\code{character}]: Names of the features.}

\item{\code{formula}}{[\code{character}]: Formula describing the learning task.}

\item{\code{target}}{[\code{character(1)}]: Name of the target column.}
}}

\examples{
task = TaskClassif$new("iris", data = iris, target = "Species")
task$formula
}
\seealso{
Other Tasks: \code{\link{TaskClassif}},
  \code{\link{TaskRegr}}
}
\keyword{datasets}
