% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif.R
\docType{data}
\name{TaskClassif}
\alias{TaskClassif}
\title{Classification Tasks}
\format{\code{\link{R6Class}} object}
\usage{
TaskClassif
}
\value{
[\code{\link{TaskClassif}}].
}
\description{
A \code{\link[R6]{R6Class}} to construct classification tasks.
}
\section{Fields}{

\describe{
\item{\code{id}}{[\code{character(1)}]: Identifier of the task.}

\item{\code{features}}{[\code{character}]: Names of the features.}

\item{\code{formula}}{[\code{character}]: Formula describing the learning task.}

\item{\code{target}}{[\code{character(1)}]: Name of the target column.}

\item{\code{positive}}{[\code{character(1)}]: Only for binary classification: Level of the positive class (\code{NA} otherwise).}

\item{\code{classes}}{[\code{character()}]: Levels of class labels.}

\item{\code{nclasses}}{[\code{integer(1)}]: Number of levels of class labels.}
}}

\examples{
task = TaskClassif$new("iris", data = iris, target = "Species")
task$formula
}
\seealso{
Other Tasks: \code{\link{TaskRegr}},
  \code{\link{TaskSupervised}}, \code{\link{Task}}
}
\keyword{datasets}
