% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr.R
\docType{data}
\name{TaskRegr}
\alias{TaskRegr}
\title{Regression Tasks}
\format{\code{\link[=R6Class]{R6Class()}} object}
\usage{
TaskRegr
}
\value{
\code{\link[=TaskRegr]{TaskRegr()}}.
}
\description{
A \code{\link[R6:R6Class]{R6::R6Class()}} to construct regression tasks.
}
\section{Fields}{

\describe{
\item{\code{id}}{(`character(1)`)\cr
Identifier of the task.}

\item{\code{backend}}{(`Backend`)\cr
Internal abstraction for data access. Leave this alone unless you know exactly what you are doing.}

\item{\code{get}}{(`function(rows = NULL, cols = NULL)`)\cr
Get (a subset of) the task data. Rows must be specified by row ids (defaulting to integers if not specified otherwise), columns must be addressed by name.}

\item{\code{head}}{(`function(n = 6L)`)\cr
Fetches the first \code{n} rows of the data and returns a \code{data.table}.}

\item{\code{features}}{[character]\cr
Names of the features.}

\item{\code{formula}}{[character]\cr
Formula describing the learning task.}

\item{\code{subset}}{(`function(rows = NULL, cols = NULL)`)\cr
Subset this task to contain only the specified rows and columns.}

\item{\code{nrow}}{(`integer(1)`)\cr
Number of rows in the current view of the task.}

\item{\code{ncol}}{(`integer(1)`)\cr
Number of columns in the current view of the task.}

\item{\code{col.types}}{(`named character`)\cr
Column class information.}

\item{\code{missing.values}}{(`named integer`)\cr
Number of missing values per column.}

\item{\code{target}}{(`character(1)`)\cr
Name of the target column.}

\item{\code{truth}}{(`function(rows = NULL`)\cr
Returns \code{data.table} with the true outcome for specified rows.}
}}

\examples{
task = TaskRegr$new(data = iris, target = "Sepal.Length")
task$formula
}
\seealso{
Other Tasks: \code{\link{TaskClassif}}, \code{\link{Task}}
}
\keyword{datasets}
