% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainLog.R
\docType{data}
\name{TrainLog}
\alias{TrainLog}
\title{TrainLog}
\format{\code{\link{R6Class}} object}
\usage{
TrainLog
}
\description{
A \code{\link[R6]{R6Class}} containing logging information of a \code{\link{train}} call.
Constructor is called with output [\code{list}]: List of output, messages, warnings and errors occured in training and the training time.
}
\section{Fields}{

\describe{
\item{\code{train.time}}{[\code{numeric(1)}]: Training time.}

\item{\code{messages}}{[\code{list}]: List of messages encountered while training the model.}

\item{\code{warnings}}{[\code{list}]: List of warnings encountered while training the model.}

\item{\code{errors}}{[\code{list}]: List of errors encountered while training the model.}

\item{\code{output}}{[\code{list}]: Remaining output from model fit, only saved if \code{mlrng.keep.train.output} option is set to \code{TRUE}.}
}}

\keyword{datasets}
